pyproject.toml
[project]
name = "rag-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.13",
    "groq>=0.28.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "pypdf>=5.6.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
    "uvicorn>=0.34.3",
]
